feat: create BlockedToken model and move User model
