+++ backend/app/features/auth/models.py
@@ -1,12 +1,10 @@
-from sqlalchemy import String, Integer
+from datetime import datetime
+from sqlalchemy import String, DateTime
 from sqlalchemy.orm import Mapped, mapped_column
 from app.database.base import Base
-
-
-class User(Base):
-    __tablename__ = 'users'
     
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    login_id: Mapped[str] = mapped_column(String(50), unique=True, nullable=False)
-    hashed_password: Mapped[str] = mapped_column(String(100), nullable=False)
-    username: Mapped[str] = mapped_column(String(100), nullable=True)
+class BlockedToken(Base):
+    __tablename__ = 'blocked_tokens'
+    
+    token_id: Mapped[str] = mapped_column(String(255), primary_key=True)
+    expired_at: Mapped[datetime] = mapped_column(DateTime)
